requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.12.0
requests==2.24.0
typing==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1

pytest~=6.1.0
selenium~=3.141.0